backend/pyproject.toml
[project]
name = "rxdx-backend"
version = "0.1.0"
description = "RxDx - Project Management System for Regulated Industries"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
    "aiosmtplib>=3.0.1",
    "aiohttp>=3.9.1",
    "cryptography>=42.0.0",
    "reportlab>=4.0.9",
    "openpyxl>=3.1.2",
    "docxtpl>=0.16.7",
    "python-docx>=1.1.0",
    "ortools>=9.8.3296",
    "passlib[argon2]>=1.7.4",
    "httpx>=0.28.1",
    "hypothesis>=6.150.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "aioimaplib>=2.0.1",
    "jsonschema>=4.26.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.96.1",
    "httpx>=0.26.0",
    "ruff>=0.1.14",
    "black>=24.1.1",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning",
    "ignore:Accessing argon2.__version__:DeprecationWarning",
    "ignore:datetime.datetime.utcnow\\(\\) is deprecated:DeprecationWarning",
    "ignore:'HTTP_422_UNPROCESSABLE_ENTITY' is deprecated:DeprecationWarning",
]

[tool.ruff]
line-length = 88
target-version = "py311"
lint.select = ["E", "F", "I", "N", "W", "UP"]
lint.ignore = ["E501"]
lint.extend-safe-fixes = ["W293"]        # Blank line contains whitespace

[tool.ruff.lint.per-file-ignores]
# Ignore E712 in all files under the tests directory
"tests/**/*.py" = ["E712", "F811", "F841"]
# ignore here too
"app/services/signature_service.py" = ["E712"]

[tool.ty.environment]
root = ["./app"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "hypothesis>=6.150.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
